Report 상태 처리완료 시 피신고자 reportCount 증가 기능 완료

+++ TeamProject3-MentoringSite/src/main/java/com/itwill/jpa/dto/report/ReportDto.java
@@ -17,7 +17,7 @@ import lombok.NoArgsConstructor;
 public class ReportDto {
 	private Long reportNo;
 	private String reportType;
-	private Integer reportTarget;
+	private Long reportTarget;
 	private Integer reportReason;
 	private String reportContent;
 	private LocalDateTime reportDate;
